requirements.txt
pandas==1.3.4
ta==0.10.2
python-telegram-bot==13.4
requests==2.31.0
yfinance==0.2.30